test(sessions): move append media coverage to transcript owner
