+++ src/config/sessions/session-accessor.sqlite-transcript-store.test.ts
@@ -1,0 +1,63 @@
+import { afterEach, describe, expect, it } from "vitest";
+import { cleanupTempDirs, makeTempDir } from "../../../test/helpers/temp-dir.js";
+import {
+  closeOpenClawAgentDatabasesForTest,
+  openOpenClawAgentDatabase,
+  runOpenClawAgentWriteTransaction,
+} from "../../state/openclaw-agent-db.js";
+import { closeOpenClawStateDatabaseForTest } from "../../state/openclaw-state-db.js";
+import { appendTranscriptEventInTransaction } from "./session-accessor.sqlite-transcript-store.js";
+
+const tempDirs: string[] = [];
+
+afterEach(() => {
+  closeOpenClawAgentDatabasesForTest();
+  closeOpenClawStateDatabaseForTest();
+  cleanupTempDirs(tempDirs);
+});
+
+describe("SQLite transcript append", () => {
+  it("canonicalizes assistant media at the generic transcript append owner", async () => {
+    const stateDir = makeTempDir(tempDirs, "media-persistence-append-");
+    const env = { OPENCLAW_STATE_DIR: stateDir };
+    runOpenClawAgentWriteTransaction(
+      (database) => {
+        expect(
+          appendTranscriptEventInTransaction(
+            database,
+            {
+              agentId: "main",
+              env,
+              sessionId: "append-session",
+              sessionKey: "agent:main:append-session",
+            },
+            {
+              type: "message",
+              id: "event-1",
+              parentId: null,
+              timestamp: 1000,
+              message: {
+                role: "assistant",
+                content: "append",
+                MediaPaths: ["/media/a.png"],
+                MediaTypes: ["image/png"],
+              },
+            },
+          ),
+        ).toBe(true);
+      },
+      { agentId: "main", env },
+    );
+    const database = openOpenClawAgentDatabase({ agentId: "main", env });
+    const row = database.db
+      .prepare("SELECT event_json FROM transcript_events WHERE session_id = ? AND seq = 0")
+      .get("append-session") as { event_json: string };
+    const message = (JSON.parse(row.event_json) as { message: Record<string, unknown> }).message;
+    expect(message).toMatchObject({ role: "assistant", content: "append" });
+    expect(message).not.toHaveProperty("MediaPaths");
+    expect(message).not.toHaveProperty("MediaTypes");
+    expect(message["__openclaw"]).toMatchObject({
+      media: [expect.objectContaining({ path: "/media/a.png", contentType: "image/png" })],
+    });
+  });
+});
